updated version, add correction
